Namespace merge to main branch

+++ twmap_gen/lib/twmapdb.inc.php
@@ -202,9 +202,9 @@ function map_get_ids($uid, $limit = 10) {
 	return $rs;
 }
 // 取 ok, expired  flag = 0 or 1 的地圖, 用來算限制
-function map_list_get($uid) {
+function map_list_get($uid, $order='') {
 	$db=get_conn();
-	$sql = sprintf("select * from \"map\" where \"uid\"=%s AND (flag = 1 or flag = 0) ORDER BY mid",$uid);
+	$sql = sprintf("select * from \"map\" where \"uid\"=%s AND (flag = 1 or flag = 0) ORDER BY mid %s",$uid,$order);
 	$rs = $db->GetAll($sql);
 	$db->close();
 	logsql($sql,$rs);
@@ -452,6 +452,8 @@ function map_expire($mid) {
 	foreach($files as $f) {
 		// 不刪除 gpx 檔案
 		if(strstr(basename($f),'.gpx')) continue;
+		// 不刪除 cmd 跟 txt for reference
+		if(strstr(basename($f),'.txt') || strstr(basename($f),".cmd")) continue;
 		$ret = unlink($f);
 		if ($ret === false ) {
 			return false;
